some minor renaming/reorganizing
